Java OOP Advanced/Open-Closed-Liskov-Substitution/src/skeleton/P05_Square/RectangleTests.java
package skeleton.P05_Square;


import org.junit.Assert;
import org.junit.Test;

public class RectangleTests {
    @Test
    public void getSidesTest() {
        Rectangle rect = new Rectangle();
        rect.setHeight(5);
        rect.setWidth(10);
        Assert.assertEquals(10, rect.getWidth());
        Assert.assertEquals(5, rect.getHeight());
    }

    @Test
    public void getAreaTest() {
        Rectangle rect = new Rectangle();
        rect.setHeight(5);
        rect.setWidth(10);
        Assert.assertEquals(50, rect.getArea());
    }

    @Test(expected = IllegalArgumentException.class)
    public void getSidesTestForSquare() {
        Rectangle rect = new Square(5, 10);
    }
}
